Fixed PIN authentication bypass
Due to some previous card operation the key may be unlocked (and ready to sign) even though the PIN was not verified in some new session object. This change enforces the PIN verification independant from the PIN's vefification status.

Before this change, a PIN verification may be bypassed, for example, in the following use case:
	sudo sc_auth pair   -h 7B9E7F52F0309C57CF4C252A5D30B917880DE46F -u username
	sudo sc_auth unpair -h 7B9E7F52F0309C57CF4C252A5D30B917880DE46F -u username
	sudo sc_auth pair   -h 7B9E7F52F0309C57CF4C252A5D30B917880DE46F -u username
In the above workflow, the PIN is only requested in the first pairing, but not in the second call. Even more severe, passing an empty PIN to the Login screen would unlock the screen, since this only checks the PIN verification status ("verified in a previous session") and then uses the unlocked key to sign the login challenge.

With this change, OpenSC now enforces a logout of the PIN if the session is deleted. Some cards, however, are not capable of explicitly logging out, so we additionally keep track of the authentication state to enforce the PIN verification in software even though the key would technically be already unlocked.